chore(todo): move todo components into seperate folder
